feat: separate out Encrypt from Enigma initial commit
